updated user bill id array

+++ app/api/bills/[id]/delete/route.ts
@@ -5,16 +5,45 @@ import {
 import prisma from "@/prisma/db";
 import { Prisma } from "@prisma/client";
 
+interface RequestBody {
+  email: string;
+}
+
 export async function DELETE(
   req: Request,
   { params }: { params: { id: string } }
 ) {
+  if (!req.body) {
+    return apiErrorResponse(400, "No data provided");
+  }
+
   try {
+    const body: RequestBody = await req.json();
+    const { email } = body;
     const { id } = params;
 
     await prisma.bill.delete({
       where: { id },
     });
+
+    const user = await prisma.user.findUnique({
+      where: { email: email },
+      select: { billIds: true },
+    });
+
+    if (user && user.billIds) {
+      const updatedBillIds = user.billIds.filter((billId) => billId !== id);
+
+      await prisma.user.update({
+        where: { email: email },
+        data: {
+          billIds: {
+            set: updatedBillIds,
+          },
+        },
+      });
+    }
+
     return apiSuccessResponse(200, {
       message: "Bill deleted successfully",
     });
